MSI: 502: Create InventoryConfiguration module, add GetStockItemDataTest
